pyproject.toml
[tool.poetry]
name = "up-bank-plot"
version = "0.1.0"
description = "A project to analyze and plot data"
authors = ["y"]
license = "y"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
plotly = "^5.24.1"
networkx = "^3.4.2"
black = "^24.10.0"
pytest = "^8.3.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
